# web/public/openapi.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "Machina API",
    "version": "1.0.0",
    "description": "REST API for Linux hypervisor hosts: QEMU/KVM domains via libvirt, networks, storage, host networking, backups, metrics, and optional KubeVirt bundle/apply workflows when configured."
  },
  "servers": [
    { "url": "/api/v1", "description": "Default API base" }
  ],
  "security": [
    { "bearerAuth": [] },
    { "cookieAuth": [] }
  ],
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "description": "Bearer token authentication"
      },
      "cookieAuth": {
        "type": "apiKey",
        "in": "cookie",
        "name": "machina_session",
        "description": "Cookie-based session authentication (`machina_session=`)"
      }
    },
    "schemas": {
...
      "SnapshotInfo": {
        "type": "object",
        "properties": {
          "name": { "type": "string" },
          "description": { "type": "string" },
          "state": { "type": "string" },
          "creation_time": { "type": "integer" },
          "is_current": { "type": "boolean" }
        }
      },
...
      "StoragePool": {
        "type": "object",
        "properties": {
          "name": { "type": "string" },
          "state": { "type": "string" },
          "autostart": { "type": "boolean" },
          "capacity_bytes": { "type": "integer" },
          "allocation_bytes": { "type": "integer" },
          "available_bytes": { "type": "integer" }
        }
      },
...
    }
  },
  "paths": {
...
    "/vms/{vm_name}/snapshots": {
      "get": {
        "summary": "List snapshots for a specific VM",
        "tags": ["Snapshots"],
        "parameters": [{ "name": "vm_name", "in": "path", "required": true, "schema": { "type": "string" } }],
        "responses": { "200": { "description": "VM snapshots", "content": { "application/json": { "schema": { "type": "array", "items": { "$ref": "#/components/schemas/SnapshotInfo" } } } } } }
      },
      "post": {
        "summary": "Create a snapshot for a VM",
        "tags": ["Snapshots"],
        "parameters": [{ "name": "vm_name", "in": "path", "required": true, "schema": { "type": "string" } }],
        "requestBody": { "content": { "application/json": { "schema": { "type": "object", "properties": { "name": { "type": "string" }, "description": { "type": "string" } } } } } },
        "responses": { "200": { "description": "Snapshot created" } }
      }
    },
    "/vms/{vm_name}/snapshots/{snap_name}": {
      "delete": {
        "summary": "Delete a snapshot",
        "tags": ["Snapshots"],
        "parameters": [
          { "name": "vm_name", "in": "path", "required": true, "schema": { "type": "string" } },
          { "name": "snap_name", "in": "path", "required": true, "schema": { "type": "string" } }
        ],
        "responses": { "200": { "description": "Snapshot deleted" } }
      }
    },
    "/vms/{vm_name}/snapshots/{snap_name}/revert": {
      "post": {
        "summary": "Revert VM to a snapshot",
        "tags": ["Snapshots"],
        "parameters": [
          { "name": "vm_name", "in": "path", "required": true, "schema": { "type": "string" } },
          { "name": "snap_name", "in": "path", "required": true, "schema": { "type": "string" } }
        ],
        "responses": { "200": { "description": "Reverted to snapshot" } }
      }
    },
...
    "/storage/pools": {
      "get": {
        "summary": "List all storage pools",
        "tags": ["Storage"],
        "responses": { "200": { "description": "Storage pool list", "content": { "application/json": { "schema": { "type": "array", "items": { "$ref": "#/components/schemas/StoragePool" } } } } } }
      },
      "post": {
        "summary": "Create a storage pool",
        "tags": ["Storage"],
        "requestBody": { "content": { "application/json": { "schema": { "type": "object", "properties": { "name": { "type": "string" }, "pool_type": { "type": "string" }, "target_path": { "type": "string" } } } } } },
        "responses": { "200": { "description": "Pool created" } }
      }
    },
    "/storage/pools/{name}/start": {
      "post": {
        "summary": "Start a storage pool",
        "tags": ["Storage"],
        "parameters": [{ "name": "name", "in": "path", "required": true, "schema": { "type": "string" } }],
        "responses": { "200": { "description": "Pool started" } }
      }
    },
    "/storage/pools/{name}/stop": {
      "post": {
        "summary": "Stop a storage pool",
        "tags": ["Storage"],
        "parameters": [{ "name": "name", "in": "path", "required": true, "schema": { "type": "string" } }],
        "responses": { "200": { "description": "Pool stopped" } }
      }
    },
    "/storage/pools/{name}/refresh": {
      "post": {
        "summary": "Refresh a storage pool",
        "tags": ["Storage"],
        "parameters": [{ "name": "name", "in": "path", "required": true, "schema": { "type": "string" } }],
        "responses": { "200": { "description": "Pool refreshed" } }
      }
    },
    "/storage/pools/{name}/autostart/{enabled}": {
      "post": {
        "summary": "Set storage pool autostart",
        "tags": ["Storage"],
        "parameters": [
          { "name": "name", "in": "path", "required": true, "schema": { "type": "string" } },
          { "name": "enabled", "in": "path", "required": true, "schema": { "type": "boolean" } }
        ],
        "responses": { "200": { "description": "Pool autostart updated" } }
      }
    },
...
    "/storage/pools/{pool_name}/volumes": {
      "get": {
        "summary": "List volumes in a storage pool",
        "tags": ["Storage"],
        "parameters": [{ "name": "pool_name", "in": "path", "required": true, "schema": { "type": "string" } }],
        "responses": { "200": { "description": "Volume list" } }
      },
      "post": {
        "summary": "Create a volume in a storage pool",
        "tags": ["Storage"],
        "parameters": [{ "name": "pool_name", "in": "path", "required": true, "schema": { "type": "string" } }],
        "requestBody": { "content": { "application/json": { "schema": { "type": "object", "properties": { "name": { "type": "string" }, "capacity_gb": { "type": "number" }, "format": { "type": "string" } } } } } },
        "responses": { "200": { "description": "Volume created" } }
      }
    },
    "/storage/pools/{pool_name}/volumes/{vol_name}": {
      "delete": {
        "summary": "Delete a volume",
        "tags": ["Storage"],
        "parameters": [
          { "name": "pool_name", "in": "path", "required": true, "schema": { "type": "string" } },
          { "name": "vol_name", "in": "path", "required": true, "schema": { "type": "string" } }
        ],
        "responses": { "200": { "description": "Volume deleted" } }
      }
    },
...
  }
}